Version 5 update

+++ mitm_request/main/request.c
@@ -211,14 +211,12 @@ esp_err_t _http_event_handler(esp_http_client_event_t *evt)
           }
           output_len = 0;
           break;
-/* This does not work (exist) in esp-idf version 4.4
       case HTTP_EVENT_REDIRECT:
           ESP_LOGD(TAG, "HTTP_EVENT_REDIRECT");
           esp_http_client_set_header(evt->client, "From", "user@example.com");
           esp_http_client_set_header(evt->client, "Accept", "text/html");
           esp_http_client_set_redirection(evt->client);
           break;
-*/
   }
   return ESP_OK;
 }
@@ -274,8 +272,8 @@ void http_request(void* args) {
     esp_err_t err = esp_http_client_perform(client);
     if (err == ESP_OK) {
         ESP_LOGI(TAG, "HTTP GET Status = %d, content_length = %d",
-                esp_http_client_get_status_code(client),
-                esp_http_client_get_content_length(client));
+                (int)esp_http_client_get_status_code(client),
+                (int)esp_http_client_get_content_length(client));
     } else {
         ESP_LOGE(TAG, "HTTP GET request failed: %s", esp_err_to_name(err));
     }
